Mod: Tests for `LaneInfo.js` also moved
